menu bugfix

+++ install.sh
@@ -5,8 +5,6 @@
 Color_Off='\033[0m'       # Text Reset
 On_Red='\033[41m'         # Red
 
-t=$(mktemp);
-
 
 BASEDIR=$(dirname "$0")
 USERNAME="$SUDO_USER"
@@ -36,43 +34,41 @@ app_install(){
     app=$1
     if command -v "nala" &> /dev/null
     then
-        #whiptail --textbox /dev/stdin 40 80 <<<"$(sudo nala install "$app" -y)"
-        # sudo nala install "$app" -y > $t; whiptail --msgbox $t 40 80
         sudo nala install "$app" -y
     else
         sudo apt -y install "$app"
     fi
+if \[ "\$1" -eq 0 \]; then
+    if [ "$2" -eq 1 ]; then
+        menu
+    fi
 }
 
 wego() {
-    # {
-    #     app_install golang-go
-    #     echo 30
-    #     export GOPATH=/home/"$USERNAME"/gocode #test it sudo?
-    #     echo 35
-    #     go install github.com/schachmat/wego@latest
-    #     echo 45
-    #     /home/"$USERNAME"/gocode/bin/wego
-    #     echo 50
-    #     API=$(whiptail --title "OpenWeather API" --inputbox "Please insert your OW API key?" 10 60 3>&1 1>&2 2>&3)
-    #     sudo sed -i 's/location=40.748,-73.985/location=Dorog/' /home/"$USERNAME"/.wegorc
-    #     sudo sed -i "s/owm-api-key=/owm-api-key=$API/" /home/"$USERNAME"/.wegorc
-    #     sudo sed -i 's/backend=forecast.io/backend=openweathermap/' /home/"$USERNAME"/.wegorc
-    #     sudo sed -i 's/owm-lang=en/owm-lang=hu/' /home/"$USERNAME"/.wegorc
-    #     echo 100
-    #     /home/"$USERNAME"/gocode/bin/wego
-    # } | whiptail --gauge "Please wait while installing" 6 60 0
-    app_install golang-go
-    export GOPATH=/home/"$USERNAME"/gocode #test it sudo?
-    go install github.com/schachmat/wego@latest
-    /home/"$USERNAME"/gocode/bin/wego
-    API=$(whiptail --title "OpenWeather API" --inputbox "Please insert your OW API key?" 10 60 3>&1 1>&2 2>&3)
-    sudo sed -i 's/location=40.748,-73.985/location=Dorog/' /home/"$USERNAME"/.wegorc
-    sudo sed -i "s/owm-api-key=/owm-api-key=$API/" /home/"$USERNAME"/.wegorc
-    sudo sed -i 's/backend=forecast.io/backend=openweathermap/' /home/"$USERNAME"/.wegorc
-    sudo sed -i 's/owm-lang=en/owm-lang=hu/' /home/"$USERNAME"/.wegorc
-    /home/"$USERNAME"/gocode/bin/wego
-    menu
+    {
+        app_install golang-go 0
+        echo 30
+        export GOPATH=/home/"$USERNAME"/gocode #test it sudo?
+        echo 35
+        go install github.com/schachmat/wego@latest
+        echo 45
+        /home/"$USERNAME"/gocode/bin/wego
+        echo 50
+        API=$(whiptail --title "OpenWeather API" --inputbox "Please insert your OW API key?" 10 60 3>&1 1>&2 2>&3)
+        sudo sed -i 's/location=40.748,-73.985/location=Dorog/' /home/"$USERNAME"/.wegorc
+        sudo sed -i "s/owm-api-key=/owm-api-key=$API/" /home/"$USERNAME"/.wegorc
+        sudo sed -i 's/backend=forecast.io/backend=openweathermap/' /home/"$USERNAME"/.wegorc
+        sudo sed -i 's/owm-lang=en/owm-lang=hu/' /home/"$USERNAME"/.wegorc
+        echo 100
+        /home/"$USERNAME"/gocode/bin/wego
+
+    } | whiptail --gauge "Please wait while installing" 6 60 0
+    if [ "$1" -eq 1 ]; then
+        menu
+    fi
+  
+  
+
 }
 
 update_upgrade(){
@@ -115,7 +111,9 @@ nala_install(){
     wget -qO - https://deb.volian.org/volian/scar.key | sudo tee /etc/apt/trusted.gpg.d/volian-archive-scar-unstable.gpg
     sudo apt update 
     sudo apt -y install nala
-    menu
+    if [ "$1" -eq 1 ]; then
+        menu
+    fi
 }
 
 startup(){
@@ -145,18 +143,22 @@ startup(){
         echo    # (optional) move to a new line
         if [[ $REPLY =~ ^[Yy]$ ]]
         then
-            install_neofetch
+            neofetch
             sudo echo 'neofetch' >> /etc/profile.d/startup.sh
         fi
     else
         sudo echo 'neofetch' >> /etc/profile.d/startup.sh
     fi
-    menu
+    if [ "$1" -eq 1 ]; then
+        menu
+    fi
 }
 
-supervisor(){
-    app_install supervisor
-    menu
+supervisor_install(){
+    app_install supervisor 0
+    if [ "$1" -eq 1 ]; then
+        menu
+    fi
 }
 
 progress(){
@@ -187,37 +189,6 @@ progress(){
     #   "${#array_two[@]}" "${array_two[@]}"   
 }
 
-# old method
-autoinstall(){
-    msgs=("Installing nala..."
-        "Installing wego..."
-        "Installing neofetch..."
-        "Installing supervisor..."
-        "Configurate startup script..."
-        )
-    commands=("nala_install"
-            "wego"
-            "install_neofetch"
-            "supervisor"
-            "startup"
-            )
-
-    n=${#commands[@]}
-    i=0
-    while [ "$i" -lt "$n" ]; do
-        pct=$(( i * 100 / n ))
-        echo XXX
-        echo $i
-        echo "${msgs[i]}"
-        echo XXX
-        echo "$pct"
-        eval "${commands[i]}"
-        i=$((i + 1))
-        sleep 1
-    done | whiptail --title "Auto install" --gauge "Preparing install..." 10 60 0
-    menu
-}
-
 reboot(){
     sudo reboot
 }
@@ -232,13 +203,14 @@ edge(){
     sudo sh -c 'echo "deb [arch=amd64] https://packages.microsoft.com/repos/edge stable main" > /etc/apt/sources.list.d/microsoft-edge-dev.list'
     sudo rm microsoft.gpg
     update
-    app_install microsoft-edge-stable
-    menu
+    app_install microsoft-edge-stable 0
+    if [ "$1" -eq 1 ]; then
+        menu
+    fi
 }
 
-install_neofetch(){
-    app_install neofetch
-    neofetch
+neofetch_install(){
+    app_install neofetch 0
     sudo sed -i 's/disk_display="off"/disk_display="infobar"/' /home/"$USERNAME"/.config/neofetch/config.conf
     sudo sed -i 's/memory_display="off"/memory_display="infobar"/' /home/"$USERNAME"/.config/neofetch/config.conf
     sudo sed -i 's/cpu_display="off"/cpu_display="infobar"/' /home/"$USERNAME"/.config/neofetch/config.conf
@@ -254,7 +226,9 @@ install_neofetch(){
     sudo sed -i 's/\# info "Local IP" local_ip/info "Local IP" local_ip/' /home/"$USERNAME"/.config/neofetch/config.conf
     sudo sed -i 's/\# info "Disk" disk/info "Disk" disk/' /home/"$USERNAME"/.config/neofetch/config.conf
     sudo sed -i 's/info "Packages" packages/\# info "Packages" packages/' /home/"$USERNAME"/.config/neofetch/config.conf
-    menu
+    if [ "$1" -eq 1 ]; then
+        menu
+    fi
 }
 
 ascii(){
@@ -264,11 +238,10 @@ ascii(){
     menu
 }
 
-lsd(){
+lsd_install(){
     # sudo nala fonts-hack-ttf
-    # if ! command -v "lsd" &> /dev/null
-    # then
-        echo "install LSD"
+    if ! command -v "lsd" &> /dev/null
+    then
         wget https://github.com/Peltoche/lsd/releases/download/0.22.0/lsd_0.22.0_amd64.deb
         sudo dpkg -i lsd_0.22.0_amd64.deb
         rm -R lsd_0.22.0_amd64.deb
@@ -276,7 +249,7 @@ lsd(){
         then
             sudo sed -i "s/\# some more ls aliases/\# some more ls aliases\nalias ls='lsd -l'/" /home/"$USERNAME"/.bashrc
         fi
-    # fi
+    fi
 
     FILE=/home/"$USERNAME"/.config/lsd
     if test -f "$FILE"/config.yaml;
@@ -324,7 +297,9 @@ lsd(){
     fi
 
     #cd /home/"$USERNAME"/.local/share/fonts && curl -fLo "Droid Sans Mono for Powerline Nerd Font Complete.otf" https://github.com/ryanoasis/nerd-fonts/raw/master/patched-fonts/DroidSansMono/complete/Droid%20Sans%20Mono%20Nerd%20Font%20Complete.otf
-    menu
+    if [ "$1" -eq 1 ]; then
+        menu
+    fi
 
 }
 
@@ -341,18 +316,18 @@ _update(){
 
 c_autoinstall(){
     STEP_LIST=(
-        'wego' 'wego (Terminal weather)'
-        'neofetch' 'neofetch (Shows Linux System Information)'
-        'supervisor' 'Supervisor (System for controlling process state)'
-        'nala_install' 'nala (APT package manager)'
-        'edge' 'DE Microsoft Edge'
-        'app_install lolcat' 'lolcat (colored text)'
-        'app_install figlet' 'figlet (ascii art text)'
-        'app_install ncdu' 'ncdu (disk usage viewer)'
-        'app_install ranger' 'ranger (Console File Manager with VI Key Bindings)'
-        'app_install bpytop' 'bpytop (Resource monitor)'
-        'lsd' 'lsd (File manager)'
-        'startup' 'Create a startup file (neofetch and wego)'
+        'wego 0' 'wego (Terminal weather)'
+        'neofetch_install 0' 'neofetch (Shows Linux System Information)'
+        'supervisor_install 0' 'Supervisor (System for controlling process state)'
+        'nala_install 0' 'nala (APT package manager)'
+        'edge 0' 'DE Microsoft Edge'
+        'app_install lolcat 0' 'lolcat (colored text)'
+        'app_install figlet 0' 'figlet (ascii art text)'
+        'app_install ncdu 0' 'ncdu (disk usage viewer)'
+        'app_install ranger 0' 'ranger (Console File Manager with VI Key Bindings)'
+        'app_install bpytop 0' 'bpytop (Resource monitor)'
+        'lsd_install 0' 'lsd (File manager)'
+        'startup 0' 'Create a startup file (neofetch and wego)'
     )
 
     entry_options=()
@@ -414,18 +389,18 @@ menu(){
     if [ $exitstatus = 0 ]; then
         case $OPTION in
             1) c_autoinstall;;
-            2) nala_install;;
-            3) wego;;
-            4) supervisor;;
-            5) edge;;
-            6) install_neofetch;;
-            7) app_install lolcat; menu;;
-            8) app_install figlet; menu;;
-            9) app_install ncdu; menu;;
-            10) app_install ranger; menu;;
-            11) app_install bpytop; menu;;
-            12) lsd;;
-            13) startup;;
+            2) nala_install 1;;
+            3) wego 1;;
+            4) supervisor_install 1;;
+            5) edge 1;;
+            6) neofetch_install 1;;
+            7) app_install lolcat 1;;
+            8) app_install figlet 1;;
+            9) app_install ncdu 1;;
+            10) app_install ranger 1;;
+            11) app_install bpytop 1;;
+            12) lsd_install 1;;
+            13) startup 1;;
             14) update_upgrade;;
             15) _update;;
         esac
@@ -437,5 +412,6 @@ menu(){
 
 dialog_install
 menu
+
